requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.1
appnope==0.1.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
blas==2.116
blas-devel==3.9.0
blinker==1.5
bottleneck==1.3.5
brotli==1.0.9
brotli-bin==1.0.9
brotlipy==0.7.0
bzip2==1.0.8
c-ares==1.18.1
ca-certificates==2022.12.7
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmaes==0.9.1
colorama==0.4.6
colorlog==5.0.1
comm==0.1.2
contourpy==1.0.5
cryptography==38.0.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fftw==3.3.9
flask==2.1.3
flit-core==3.6.0
fonttools==4.25.0
freetype==2.10.4
frozenlist==1.3.3
fsspec==2023.1.0
gettext==0.21.0
giflib==5.2.1
gmp==6.2.1
gnutls==3.6.15
google-auth==2.16.0
google-auth-oauthlib==0.4.6
greenlet==2.0.1
gunicorn==20.1.0
icu==68.1
idna==3.4
importlib-metadata==6.0.0
importlib_resources==5.10.2
ipykernel==6.19.2
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.1
jinja2==3.1.2
joblib==1.1.1
jpeg==9e
jupyter_client==7.4.8
jupyter_core==5.1.1
kiwisolver==1.4.4
lame==3.100
lcms2==2.14
lerc==3.0
libblas==3.9.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcblas==3.9.0
libcxx==14.0.6
libdeflate==1.8
libffi==3.4.2
libgfortran==5.0.0
libgfortran5==11.3.0
libiconv==1.16
libidn2==2.3.1
liblapack==3.9.0
liblapacke==3.9.0
libopenblas==0.3.21
libopus==1.3
libpng==1.6.37
libprotobuf==3.20.1
libsodium==1.0.18
libsqlite==3.40.0
libtasn1==4.16.0
libtiff==4.4.0
libunistring==0.9.10
libvpx==1.10.0
libwebp==1.2.4
libwebp-base==1.2.4
libxml2==2.9.14
libzlib==1.2.13
lightgbm==3.3.5
lightning-utilities==0.6.0
llvm-openmp==14.0.6
lz4-c==1.9.4
mako==1.2.4
markdown==3.4.1
markupsafe==2.1.1
matplotlib-base==3.6.2
matplotlib-inline==0.1.6
multidict==6.0.2
munkres==1.1.4
natsort==8.2.0
ncurses==6.3
nest-asyncio==1.5.6
nettle==3.7.3
numexpr==2.8.4
numpy==1.23.5
numpy-base==1.23.5
oauthlib==3.2.2
openblas==0.3.21
openh264==1.8.0
openssl==3.0.8
optuna==3.1.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.3.0
pip==22.3.1
platformdirs==2.5.2
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pygments==2.11.2
pyjwt==2.6.0
pyopenssl==23.0.0
pyparsing==3.0.9
pysocks==1.7.1
python==3.9.16
python-dateutil==2.8.2
python_abi==3.9
pytorch==1.13.1
pytorch-lightning==1.9.0
pytz==2022.7
pyu2f==0.1.5
pyyaml==6.0
pyzmq==23.2.0
readline==8.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
setuptools==65.6.3
six==1.16.0
sqlalchemy==1.4.39
sqlite==3.40.1
stack_data==0.2.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==2.2.0
tk==8.6.12
torchmetrics==0.11.0
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typing-extensions==4.4.0
typing_extensions==4.4.0
tzdata==2022g
urllib3==1.26.14
wcwidth==0.2.5
werkzeug==2.1.2
wheel==0.37.1
wtforms==3.0.1
x264==1!152.20180806
xz==5.2.8
yaml==0.2.5
yarl==1.8.1
zeromq==4.3.4
zipp==3.11.0
zlib==1.2.13
zstd==1.5.2
